+++ src/main/java/com/projet/gestionJeux/controllers/ThemeJeuController.java
@@ -4,13 +4,12 @@ import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
 
-import com.projet.gestionJeux.models.GenreJeu;
 import com.projet.gestionJeux.models.ThemeJeu;
 import com.projet.gestionJeux.services.ThemeJeuService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 
 
@@ -44,7 +43,14 @@ public class ThemeJeuController {
 		System.out.println(themeNom);
 		ThemeJeu theme = new ThemeJeu();
 		theme.setNom_theme(themeNom);
-		this.themeJeuService.saveOrUpdate(theme);
+		ThemeJeu themeJeu = this.themeJeuService.saveOrUpdate(theme);
+		
+		if (themeJeuService.existeThemeJeu(themeJeu.getId())){
+			model.put("addSucess",true);
+		}
+		else{
+			model.put("addSucess",false);
+		}
 		
 		List<ThemeJeu> themes = themeJeuService.getThemeJeux();
 		model.put("themes", themes);
@@ -60,7 +66,14 @@ public class ThemeJeuController {
 		int themeId = Integer.parseInt(theme_id);
 		ThemeJeu theme = this.themeJeuService.findById(themeId);
 		theme.setNom_theme(themeNom); 
-		this.themeJeuService.saveOrUpdate(theme);
+		ThemeJeu themeJeu = this.themeJeuService.saveOrUpdate(theme);
+		
+		if (themeJeuService.existeThemeJeu(themeJeu.getId())){
+			model.put("updateSucess",true);
+		}
+		else{
+			model.put("updateSucess",false);
+		}
 		
 		
 		List<ThemeJeu> themes = themeJeuService.getThemeJeux();
@@ -72,8 +85,19 @@ public class ThemeJeuController {
 	/**
 	 * Supprimer un thème
 	 */
-	@DeleteMapping("/themes/{themeId}")
-	public void supprimerTheme(int id) {
-		this.themeJeuService.deleteThemeJeu(id);
+	@GetMapping("/gestion-jeux/deleteTheme/{id}")
+	public String supprimerTheme(@PathVariable("id") int id, ModelMap model)
+	{
+	  if (themeJeuService.existeThemeJeu(id)) {
+	      this.themeJeuService.deleteThemeJeu(id);
+	      model.put("deleteSucess",true);
+      }
+	  else{
+		  model.put("deleteSucess",false);
+	  }
+	  
+	  List<ThemeJeu> themes = themeJeuService.getThemeJeux();
+	  model.put("themes", themes);
+	  return "Theme/GestionTheme";
 	}
 }
